data-service/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
geojson==3.2.0
haversine==2.9.0
idna==3.10
numpy==2.3.1
pika==1.3.2
psycopg2-binary==2.9.10
pyproj==3.7.1
requests==2.32.4
shapely==2.1.1
urllib3==2.5.0
